Switch to double quotes
Tobira and Editor use double quotes. Leaves only Studio, which is
kinda rewritten soon anyway, so its a good point to change it.

+++ config.js
@@ -99,7 +99,7 @@ module.exports = {
     'comma-style': 'warn',
     'curly': 'warn',
     'jsx-quotes': ['warn', 'prefer-double'],
-    'quotes': ['warn', 'single', { 'avoidEscape': true }],
+    'quotes': ['warn', 'double', { 'avoidEscape': true }],
     'semi': ['warn', 'always'],
